Move ListAcceptedAssignments and ListAllAcceptedAssignments
pkg/classroom/http.go is only meant to contain API calls.

modified:   cmd/gh-classroom/accepted-assignments/accepted_assignments.go
modified:   cmd/gh-classroom/clone/student-repos/student-repos.go
modified:   cmd/gh-classroom/shared/shared.go
modified:   pkg/classroom/http.go
modified:   pkg/classroom/http_test.go
